requirements.txt
openpyxl==3.1.2
pandas==2.2.0
selenium==4.17.2
beautifulsoup4==4.12.3
pyodbc==5.1.0
python-dotenv==1.0.1
SQLAlchemy==2.0.26
psycopg2-binary==2.9.9

